% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMagpieObject.R
\name{compareMagpieObject}
\alias{compareMagpieObject}
\title{compareMagpieObject}
\usage{
compareMagpieObject(x, y, tol = 0)
}
\arguments{
\item{x}{a magpie object}

\item{y}{a magpie object}

\item{tol}{optional value between 0 and 1 to restrict value differences between the
two magpie objects to be listed, a data entry must exceed `tol` percent of the maximum
value difference observed in the data to be listed}
}
\description{
A helper to analyze two magpie objects for differences.
}
\author{
Falk Benke
}
