% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageLocation.R
\name{getPackageLocation}
\alias{getPackageLocation}
\title{Find the path to R System Files for a given R library}
\usage{
getPackageLocation(packageName)
}
\arguments{
\item{packageName}{a character string with the name of a single package}
}
\description{
A wrapper around \link[base]{system.file} to make sure that the correct path
to the package on the local machine is returned. Built for stability when
working with devtools::load_all() during development.
}
