% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGdxToDf.R
\name{readGdxToDf}
\alias{readGdxToDf}
\title{readGdxToDf}
\usage{
readGdxToDf(symbolName, pathToGdx, ...)
}
\arguments{
\item{symbolName}{The name of a symbol to be extracted from gdx.}

\item{pathToGdx}{Path to a gdx file.}

\item{...}{Additional arguments passed to gamstransfer::readGDX.}
}
\value{
A data frame with data about symbolName from the given gdx file.
}
\description{
Read data for a given symbol from a gdx file and return it as a dataframe using
gamstransfer::readGDX
}
\author{
Falk Benke
}
