% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemFile.R
\name{getSystemFile}
\alias{getSystemFile}
\title{Find the path to R System Files for a given R library}
\usage{
getSystemFile(..., package = "base", mustWork = FALSE)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s) within some package}

\item{package}{a character string with the name of a single package}

\item{mustWork}{if TRUE, an error is given if there are no matching files}
}
\description{
A wrapper around \link[base]{system.file} to make sure that the correct paths
are returned when working with devtools::load_all() during development.
}
