% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamutils-package.R
\docType{package}
\name{piamutils-package}
\alias{piamutils}
\alias{piamutils-package}
\title{piamutils: Utilities for the piam-verse}
\description{
This package contains utilities and helpers needed in various piam libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piamutils}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Christof Schötz
}

}
\keyword{internal}
